PPU: more accurately emulate mid-frame vertical mosaic changes
Previously when writing the MOSAIC register, we would reload the vertical
mosaic counters to zero. However this is not accurate.

The vertical counters should not change when writing that register.
They continue to increment from the current value.
Eventually they will reach the new values from the MOSAIC register and be reloaded to zero,
or they wraparound when incrementing from 15 to 16 due to the 4-bit size of the counters.