DMA: FIFO A/B should be statically assigned to DMA1/2
We used to check the DMA destination address to figure out what DMA channel
belongs to what FIFO channel. This is incorrect. The correct behaviour is that
FIFO A triggers DMA1 and FIFO B triggers DMA2.

This fixes 'Magical Vacation', which sets the source and destination DMA address to zero.